backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.4
prometheus-client==0.20.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
